llm-00-foundation/04-bert/requirements.txt
# Python 3.10.12
transformers==4.51.0
torch==2.4.0
sentencepiece==0.2.0
tqdm==4.66.5
loguru==0.7.3
pytorch-crf
# 配置文件管理
omegaconf==2.3.0
hydra-core==1.3.2